Fix file rotation detection with slow WatchService implementations
Platforms which do not have a fast WatchService implementation fall back
to the PollingWatchService. This includes Mac OS X and AIX.

The PollingWatchService is pretty slow and might miss a file rename
which might happen during a log rotate. To avoid this, check the file
key of the followed file before reading from it and make sure to restart
the file follow if the key does not match anymore.